Many changes again...

+++ cogs/tournament.py
@@ -30,33 +30,20 @@ class Tournament:
 
     async def update_tournaments(self):
         self.bot.logger.info(f'Checking unconfirmed tournaments...')
-        for event in Event.select().where(not Event.confirmed):
+        for event in Event.select().where(Event.confirmed == False):
             if event.created < datetime.utcnow() - timedelta(minutes=5):
-                self.bot.logger.info(f'Deleted unconfirmed tournament with ID {event.tournament_id}!')
+                self.bot.logger.info(f'Deleted unconfirmed tournament with ID {event.guild_id}!')
                 event.delete_instance()
 
     @property
     def random_id(self):
         return str(hex(int(time() * 10000000))[8:])
 
-    async def embed_notify(self, ctx: commands.Context, error: int, title: str, message: str):
-        """Create and reply Discord embeds in one line."""
-        embed = discord.Embed()
-        embed.title = title
-        embed.colour = discord.Colour.dark_red() if error == 1 else discord.Colour.green() if error == 0 else discord.Colour.blue()
-        embed.description = message
-        embed.set_footer(text='Paragon', icon_url=self.agora.icon_url)
-        await ctx.send(embed=embed)
-
     @commands.group(pass_context=True)
     async def tournament(self, ctx):
         """Tournament commands related to this server."""
         if ctx.invoked_subcommand is None:
             await ctx.send('Please see ' + self.bot.command_prefix + 'help tournament for command usage.')
-
-    @tournament.command(name='list')
-    async def _list(self, ctx):
-        print()
 
     @tournament.command(name='create', pass_context=True)
     @commands.cooldown(1, 5, commands.BucketType.user)
@@ -65,28 +52,28 @@ class Tournament:
         embed = discord.Embed()
         try:
             Event.get(Event.creator == ctx.author.id)
-            await self.embed_notify(ctx, 1, 'Error', 'Currently you cannot create more than one event per account!')
+            await self.bot.embed_notify(ctx, 1, 'Error', 'Currently you cannot create more than one event per account!')
             return
         except DoesNotExist:
             pass
         try:
-            Event.get(Event.server_id == ctx.guild.id)
-            await self.embed_notify(ctx, 1, 'Error', 'A tournament already exists on this server!')
+            Event.get(Event.guild_id == ctx.guild.id)
+            await self.bot.embed_notify(ctx, 1, 'Error', 'A tournament already exists on this server!')
             return
         except DoesNotExist:
             pass
         if tournament_type.upper() not in self.types:
-            await self.embed_notify(ctx, 1, 'Error', 'Invalid tournament type!')
+            await self.bot.embed_notify(ctx, 1, 'Error', 'Invalid tournament type!')
             return
         try:
             tournament_time = datetime.strptime(date, '%m-%d-%Y %H:%M')
             if tournament_time > datetime.now() + timedelta(days=30):
-                await self.embed_notify(ctx, 1, 'Error',
+                await self.bot.embed_notify(ctx, 1, 'Error',
                                         'Unable to create a tournament scheduled for more than a month from now!')
         except:
-            await self.embed_notify(ctx, 1, 'Error', 'Invalid time format! Please try again.')
+            await self.bot.embed_notify(ctx, 1, 'Error', 'Invalid time format! Please try again.')
             return
-        tournament = Event(server_id=ctx.guild.id, server_name=ctx.guild.name, tournament_name=name,
+        tournament = Event(real_id=ctx.guild.id, name=ctx.guild.name, tournament_name=name,
                            type=tournament_type.upper(), event_date=tournament_time,
                            confirmed=False, created=datetime.utcnow(), creator=ctx.author.id)
         tournament.save()
@@ -108,9 +95,9 @@ class Tournament:
         """Confirm the creation of a new tournament."""
         found = False
         for tournament in Event.select():
-            if tournament.server_id == unique_id and not tournament.confirmed:
-                if ctx.author.id not in [tournament.creator, self.bot.owner.id]:
-                    await self.embed_notify(ctx, 1, 'Error',
+            if tournament.guild_id == unique_id and not tournament.confirmed:
+                if str(ctx.author.id) not in [tournament.creator, self.bot.owner_id]:
+                    await self.bot.embed_notify(ctx, 1, 'Error',
                                             'Only the event creator or server staff may confirm events!')
                     return
                 tournament.confirmed = True
@@ -118,11 +105,11 @@ class Tournament:
                 found = True
                 break
         if found:
-            await self.embed_notify(ctx, 0, 'Tournament Created',
+            await self.bot.embed_notify(ctx, 0, 'Tournament Created',
                                     'Players can now join using **' + self.bot.command_prefix + 'tournament join**')
             self.bot.logger.info('Tournament confirmed with ID ' + unique_id)
         else:
-            await self.embed_notify(ctx, 1, 'Error', 'No tournament was found with the provided ID!')
+            await self.bot.embed_notify(ctx, 1, 'Error', 'No tournament was found with the provided ID!')
 
     @tournament.command(name='help', pass_context=True)
     async def _help(self, ctx):
@@ -133,9 +120,10 @@ class Tournament:
     async def _delete(self, ctx):
         """Delete an existing tournament."""
         try:
-            tournament = Event.get(Event.server_id == ctx.guild.id)
-            if ctx.message.author.id not in [tournament.creator, self.bot.owner.id]:
-                await self.embed_notify(ctx, 1, 'Error', 'Only the event creator or server staff may delete events!')
+            tournament = Event.get(Event.guild_id == ctx.guild.id)
+            if str(ctx.author.id) not in [tournament.creator, self.bot.owner_id]:
+                await self.bot.embed_notify(ctx, 1, 'Error',
+                                            'Only the event creator or server staff may delete events!')
                 return
             for team in Team.select().where(Team.tournament == ctx.guild.id):
                 for player in Player.select().where(Player.teams.contains(team.team_id)):
@@ -146,20 +134,20 @@ class Tournament:
                     Player.tournaments.contains(ctx.guild.id)):  # Remove the tournament from players!
                 player.tournaments = str(player.tournaments).replace(ctx.guild.id + '|', '')
                 player.save()
-            await self.embed_notify(ctx, 2, 'Tournament Deleted',
+            await self.bot.embed_notify(ctx, 2, 'Tournament Deleted',
                                     'Deleted the tournament called **' + tournament.tournament_name + '**.')
             tournament.delete_instance()
             self.bot.logger.info('Tournament deleted from database')
         except DoesNotExist:
-            await self.embed_notify(ctx, 1, 'Error', 'No tournament exists on this server!')
+            await self.bot.embed_notify(ctx, 1, 'Error', 'No tournament exists on this server!')
 
     @tournament.command(name='join')
     async def _join(self, ctx):
         """Join an ongoing tournament."""
         try:
-            tournament = Event.get(Event.server_id == ctx.guild.id)
+            tournament = Event.get(Event.guild_id == ctx.guild.id)
         except DoesNotExist:
-            await self.embed_notify(ctx, 1, 'Error', 'No tournament exists on this server!')
+            await self.bot.embed_notify(ctx, 1, 'Error', 'No tournament exists on this server!')
             return
         try:
             player = Player.get(Player.discord_id == ctx.author.id)
@@ -172,32 +160,32 @@ class Tournament:
                 player.save()
             tournament.size = tournament.size + 1
             tournament.save()
-            await self.embed_notify(ctx, 0, 'Tournament Joined',
+            await self.bot.embed_notify(ctx, 0, 'Tournament Joined',
                                     'You joined **' + tournament.tournament_name + '**. Have fun!')
         except DoesNotExist:
-            await self.embed_notify(ctx, 1, 'Error', 'You must link your Epic ID to your Discord account!')
+            await self.bot.embed_notify(ctx, 1, 'Error', 'You must link your Epic ID to your Discord account!')
 
     @tournament.command(name='leave')
     async def _leave(self, ctx):
         """Leave a tournament."""
         try:
-            tournament = Event.get(Event.server_id == ctx.guild.id)
+            tournament = Event.get(Event.guild_id == ctx.guild.id)
         except DoesNotExist:
-            await self.embed_notify(ctx, 1, 'Error', 'No tournament exists on this server!')
+            await self.bot.embed_notify(ctx, 1, 'Error', 'No tournament exists on this server!')
             return
         try:
             player = Player.get(Player.discord_id == ctx.message.author.id)
             if player.tournaments is not None:
                 player.tournaments = str(player.tournaments).replace(ctx.guild.id + '|', '')
-                await self.embed_notify(ctx, 2, 'Tournament Left',
+                await self.bot.embed_notify(ctx, 2, 'Tournament Left',
                                         'You left **' + tournament.tournament_name + '**. See you later!')
                 tournament.size = tournament.size - 1
                 tournament.save()
             else:
-                await self.embed_notify(ctx, 1, 'Error', 'You are not in any tournaments!')
+                await self.bot.embed_notify(ctx, 1, 'Error', 'You are not in any tournaments!')
                 return
         except DoesNotExist:
-            await self.embed_notify(ctx, 1, 'Error', 'You must link your Epic ID to your Discord account!')
+            await self.bot.embed_notify(ctx, 1, 'Error', 'You must link your Epic ID to your Discord account!')
             return
         if player.teams is not None:
             for team in Team.select().where(Team.tournament == ctx.guild.id):
@@ -212,9 +200,9 @@ class Tournament:
         """View information about a tournament."""
         embed = discord.Embed()
         try:
-            tournament = Event.get(Event.server_id == ctx.guild.id)
+            tournament = Event.get(Event.guild_id == ctx.guild.id)
         except DoesNotExist:
-            await self.embed_notify(ctx, 1, 'Error', 'No tournament exists on this server!')
+            await self.bot.embed_notify(ctx, 1, 'Error', 'No tournament exists on this server!')
             return
         creator = discord.utils.find(lambda u: u.id == tournament.creator, self.bot.get_all_members())
         if tournament.teams is not None and len(
@@ -224,7 +212,7 @@ class Tournament:
             size = 'None'
         embed.title = tournament.tournament_name
         embed.colour = discord.Colour.blue()
-        embed.description = 'Created by ' + creator.name + ' in ' + tournament.server_name
+        embed.description = 'Created by ' + creator.name + ' in ' + tournament.guild_name
         embed.add_field(name='Created', value=tournament.created)
         embed.add_field(name='Time', value=tournament.event_date)
         embed.add_field(name='Type', value=tournament.type)
@@ -233,7 +221,7 @@ class Tournament:
         embed.set_footer(text='Paragon', icon_url=self.agora.icon_url)
         await ctx.send(embed=embed)
 
-    @tournament.group(name='team', pass_context=True)
+    @commands.group(name='team')
     async def team(self, ctx):
         """Team related commands."""
         print()
@@ -243,17 +231,17 @@ class Tournament:
     async def team_create(self, ctx, name: str):
         """Manually add a team to a tournament."""
         try:
-            tournament = Event.get(Event.server_id == ctx.guild.id)
+            tournament = Event.get(Event.guild_id == ctx.guild.id)
         except DoesNotExist:
-            await self.embed_notify(ctx, 1, 'Error', 'No tournament exists on this server!')
+            await self.bot.embed_notify(ctx, 1, 'Error', 'No tournament exists on this server!')
             return
         try:
             Team.get(Team.team_name == name)
-            await self.embed_notify(ctx, 1, 'Error', 'A team already exists with this name!')
+            await self.bot.embed_notify(ctx, 1, 'Error', 'A team already exists with this name!')
         except DoesNotExist:
             pass
         team_id = self.random_id
-        team = Team(team_id=team_id, team_name=name, tournament=tournament.server_id)
+        team = Team(team_id=team_id, team_name=name, tournament=tournament.guild_id)
         team.save()
         if tournament.teams is None:
             tournament.teams = '' + team_id + '|'
@@ -267,7 +255,7 @@ class Tournament:
         embed.add_field(name='Team Name', value=name)
         embed.add_field(name='Team ID', value=team_id)
         embed.add_field(name='Event Name', value=tournament.tournament_name, inline=False)
-        embed.add_field(name='Event ID', value=tournament.server_id)
+        embed.add_field(name='Event ID', value=tournament.guild_id)
         embed.set_footer(text='Paragon', icon_url=self.agora.icon_url)
         await ctx.send(embed=embed)
 
@@ -279,22 +267,22 @@ class Tournament:
             team = Team.get((Team.team_id == name) & (Team.tournament == ctx.guild.id))
             team_id = team.team_id
             team.delete_instance()
-            await self.embed_notify(ctx, 2, 'Team Deleted', 'Deleted ' + team.team_name)
+            await self.bot.embed_notify(ctx, 2, 'Team Deleted', 'Deleted ' + team.team_name)
             return
         except DoesNotExist:
             try:
                 team = Team.get((Team.team_name % name) & (Team.tournament == ctx.guild.id))
                 team_id = team.team_id
                 team.delete_instance()
-                await self.embed_notify(ctx, 2, 'Team Deleted', 'Deleted ' + team.team_name)
+                await self.bot.embed_notify(ctx, 2, 'Team Deleted', 'Deleted ' + team.team_name)
             except DoesNotExist:
-                await self.embed_notify(ctx, 1, 'Error', 'No team exists on this server with that name or ID!')
+                await self.bot.embed_notify(ctx, 1, 'Error', 'No team exists on this server with that name or ID!')
                 return
         try:  # Remove team from tournament
-            tournament = Event.get(Event.server_id == ctx.guild.id)
+            tournament = Event.get(Event.guild_id == ctx.guild.id)
             tournament.teams = tournament.teams.replace(team_id + '|', '')
         except DoesNotExist:
-            await self.embed_notify(ctx, 1, 'Error', 'No tournament exists on this server!')
+            await self.bot.embed_notify(ctx, 1, 'Error', 'No tournament exists on this server!')
         for player in Player.select().where(Player.teams.contains(team_id)):  # Remove team from players
             player.teams = player.teams.replace(team_id + '|', '')
             player.save()
@@ -304,9 +292,9 @@ class Tournament:
     async def team_add(self, ctx):
         """Manually add mentioned players to a team."""
         try:
-            tournament = Event.get(Event.server_id == ctx.guild.id)
+            tournament = Event.get(Event.guild_id == ctx.guild.id)
         except DoesNotExist:
-            await self.embed_notify(ctx, 1, 'Error', 'No tournament exists on this server!')
+            await self.bot.embed_notify(ctx, 1, 'Error', 'No tournament exists on this server!')
         if ctx.message.mentions is not None:
             for user in ctx.message.mentions:
                 print()
@@ -323,8 +311,8 @@ class Team(BaseModel):
 
 
 class Event(BaseModel):
-    server_id = CharField(unique=True)
-    server_name = CharField()
+    guild_id = CharField(unique=True)
+    guild_name = CharField()
     creator = CharField()
     tournament_name = CharField()
     type = CharField()
